Move ValidationEnhancer into the src/enhancers folder
